refactor: move subagent registry behind facade
